fix warnings and compile errors

+++ include/pedal/TSine.hpp
@@ -1,7 +1,7 @@
 #ifndef TSine_hpp
 #define TSine_hpp
 
-#include <math.h> //so we can use sin()
+#include <cmath> //so we can use sin()
 #include "pdlSettings.hpp"//so we can access sampleRate and bufferSize
 
 class TSine {//Pedal Trivial Sine Oscillator
@@ -31,9 +31,9 @@ class TSine {//Pedal Trivial Sine Oscillator
   inline float generateNextSample(){//return a float even if you don't use it
     //inline functions must be located in the header file, so 
     //we will define it here. 
-    currentSample = sin(phase) * amplitude;//calculate single sample
-    phase += phaseIncrement;//increment phase for the next sample
-    phase = fmod(phase, 6.283185307179586);
+    currentSample = std::sin(phase) * amplitude;//calculate single sample
+    phase += (float)phaseIncrement;//increment phase for the next sample
+    phase = fmod(phase, 6.283185307179586f);
     return currentSample;
   }
